mqtt-bridge-app/requirements.txt
paho-mqtt
azure-eventgrid
azure-core
